1. 删除保税入区不必要的字段; 2. 待申报,申报统一颜色标识; 3. 修改核放单删除删不掉的问题; 4. 核放单数据生成,展示数据的问题; 5. 将第一第二法定单位转换为代码存储; 6. 直购保税净重=毛重/1.1; 7. 入区核注清单的“清单类型”字段改为普通清单（代码为0）; 8. 入区核注清单的“报关类型代码”字段改为对应报关（代码为2）; 9. 入区核注清单表头新增“对应报关单申报单位编码”、“对应报关单申报单位社会信用代码”、”对应报关单申报单位名称“ 这三个字段 10. 核注清单所有字段改为可修改; 11. 入区核注清单表体的”备案序号“改为非必填项;

+++ src/main/java/com/xaeport/crossborder/service/bondedIEnter/EnterManifestService.java
@@ -8,6 +8,7 @@ import com.xaeport.crossborder.data.mapper.EnterManifestMapper;
 import com.xaeport.crossborder.data.status.StatusCode;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -39,6 +40,7 @@ public class EnterManifestService {
         Map<String,String> paramMap = new HashMap<String,String>();
         for (String etps_preent_no:etps_preent_nos) {
             paramMap.put("status", StatusCode.RQHFDDSB);
+            paramMap.put("zcStatus", StatusCode.RQHFDDZC);
             paramMap.put("etps_preent_no",etps_preent_no);
             paramMap.put("userId",users.getId());
             PassPortHead passPortHead = new PassPortHead();
@@ -51,18 +53,19 @@ public class EnterManifestService {
 
                 //查找关联单里的核注清单信息
                 String rtl_nos = this.enterManifestMapper.queryEnterPassportAcmp(paramMap);
-                String rtl_no = rtl_nos.replaceAll("/", ",");
-                //先恢复核注清单表头
-                paramMap.put("rtl_no",rtl_no);
-                this.enterManifestMapper.updateEnterBondInvtBsc(paramMap);
-                //再恢复核注清单表体
-                String[] etps_invt_nos = this.enterManifestMapper.queryEnterBondInvtDtID(paramMap);
-                for (String etps_invt_no:etps_invt_nos) {
-                    this.enterManifestMapper.updateEnterBondInvtDt(etps_invt_no);
+                if (!StringUtils.isEmpty(rtl_nos)) {
+                    String rtl_no = rtl_nos.replaceAll("/", ",");
+                    //先恢复核注清单表头
+                    paramMap.put("rtl_no", rtl_no);
+                    this.enterManifestMapper.updateEnterBondInvtBsc(paramMap);
+                    //再恢复核注清单表体
+                    String[] etps_invt_nos = this.enterManifestMapper.queryEnterBondInvtDtID(paramMap);
+                    for (String etps_invt_no : etps_invt_nos) {
+                        this.enterManifestMapper.updateEnterBondInvtDt(etps_invt_no);
+                    }
+                    //删除关联单信息
+                    this.enterManifestMapper.deleteEnterPassportAcmp(paramMap);
                 }
-                //删除关联单信息
-                this.enterManifestMapper.deleteEnterPassportAcmp(paramMap);
-
                 //删除核放单
                 this.enterManifestMapper.deleteEnterPassportHead(paramMap);
